renames model file from Plant to PlantModel and update imports
